refactor: remove default node/identity servers from csi-common
The default node and identity server types in csi-common were thin
wrappers that returned Unimplemented for every RPC. Our azurelustre
package already implements all required RPCs directly, so these
defaults were dead code. Remove them and move the relevant tests
into the azurelustre package.
